Update patch set 3
Patch Set 3: Code-Review-1

(4 comments)

This is a nice change, but I have some trouble understanding how it works when it is disabled. As it turns out, the unit test you wrote is failing for the disabled case, so that might add to my confusion.

Patch-set: 3
Reviewer: Gerrit User 5100 <5100@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=-1