Update patch set 1
Patch Set 1:

(5 comments)

> Patch Set 1:
> 
> Drive-by comment: It looks like this does not change PeerConnection to offer extmap-allow-mixed by default. Is that intentional?

It's mentioned in a comment in pc/sessiondescription.h that before https://bugs.webrtc.org/9712, clients could not deserialize this SDP line. I'm wondering why this line wouldn't just be ignored. Does this result in an error? Since this is the case, are you going to have to munge the SDP to enable this in the offer?

Looks good, just some small comments.

Patch-set: 1
Reviewer: Gerrit User 5800 <5800@58829da1-049c-39fb-b951-ebdcd0984767>