Update patch set 2
Patch Set 2:

(2 comments)

Added a comment about supporting this at the media level.

Also one question - Currently we respond that this is supported, meaning that we can receive mixed header extensions. Is there support for sending? If there is/when this is added I think that you'll have to incorporate some implementation that doesn't mix header extensions if the answer does not respond that it's supported.

Patch-set: 2
Reviewer: Gerrit User 5800 <5800@58829da1-049c-39fb-b951-ebdcd0984767>