Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> lgtm
> 
> Tip to make reviewing easier: upload the original patch set 1 and do the fixes as patch set 2.

Patch set 1 *is* the original here, there are no new changes.
Though in that case it's nice to specify in the commit description that it is unchanged.

Also I would recommend to use "Reland" on the original change rather than "Revert" on the revert, because it just makes a nicer commit message

Patch-set: 1
Reviewer: Gerrit User 5114 <5114@58829da1-049c-39fb-b951-ebdcd0984767>