Update patch set 1
Patch Set 1:

The tl;dr of the bug is that we crash when trying to create a stereo CN instance for a stereo G722 instance. IIUC, CN is supposed to be quietly ignored for stereo, and this is my best guess for the correct place to police that.

Patch-set: 1
Reviewer: Gerrit User 5100 <5100@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5077 <5077@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0
Work-in-progress: false