Update patch set 1
Patch Set 1:

I think it would be nice to keep it as just a scoped_refptr (not also optional). The only time we need to take action when the frame_encryptor/frame_decryptor is null is when we're clearing an existing frame_encryptor/decryptor on a sender/receiver that has a media channel active.

Patch-set: 1
Reviewer: Gerrit User 5039 <5039@58829da1-049c-39fb-b951-ebdcd0984767>