Don't reset streams for the FrameEncryptor / FrameDecryptor unless they changed.
It is possible for Audio and Video streams to get reset even when they are not
setting FrameEncryptors and FrameDecryptors on their respective RtpSenders and
RtpReceivers. The correct logic for this scenario is to only set these objects
on the underlying media channel if it is not nullptr.

One special case to this is that if a SetFrameEncryptor or SetFrameDecryptor
call is explicitly made by the client then it should force set this value on
any existing media channels regardless of whether it is nullptr or not.

This change updates the code to apply this new logic.

Bug: webrtc:9795
Change-Id: I29910b9ecc2f6f8eea371c5961ac7e9780de65d2