Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> This will become impossible after Bjorn's change in https://webrtc-review.googlesource.com/c/src/+/102622

Not impossible, but has to be done in a different way. The parser tracks the log segments so you can copy the timestamp of the first log_end and then not emit anything in NextPacket, NextAudio if the timestamp is newer than the log_end.

Patch-set: 2
Reviewer: Gerrit User 5150 <5150@58829da1-049c-39fb-b951-ebdcd0984767>