Update patch set 2
Patch Set 2:

This looks good *provided* that we can rely on those calls to not break often, and be easy to fix when they do.

An alternative, clunkier but more "legal" approach would be to define a function

  void Autoreleasepool(rtc::FunctionView<void()>);

that you'd call like this:

  // 1. code outside @autoreleasepool block
  Autoreleasepool([&]{
    // 2. code inside @autoreleasepool block
  });
  // 3. code outside @autoreleasepool block

and which you'd implement in ObjC++ something like this:

  void Autoreleasepool(rtc::FunctionView<void()> f) {
    @autoreleasepool {
      f();
    }
  }

That doesn't let you avoid using ObjC++, but it shouldn't be a problem to put it in a separate build target, right?

Patch-set: 2
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>