Update patch set 3
Patch Set 3:

(2 comments)

> Yes---since the ObjC++ and C++ code needs to be in separate translation units, the compiler will be unable to inline the Autoreleasepool() call. Easily branch-predicted non-inlined function calls are cheap, but they aren't free.

Is it not the case that constructing lambda objects in C++ is kind of expensive? It seems to me that if we went with your suggested approach we might end up allocating lambdas in a loop repeatedly. Does the compiler optimise this?

Patch-set: 3
Reviewer: Gerrit User 5520 <5520@58829da1-049c-39fb-b951-ebdcd0984767>