Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (2 comments)
> 
> > Yes---since the ObjC++ and C++ code needs to be in separate translation units, the compiler will be unable to inline the Autoreleasepool() call. Easily branch-predicted non-inlined function calls are cheap, but they aren't free.
> 
> Is it not the case that constructing lambda objects in C++ is kind of expensive? It seems to me that if we went with your suggested approach we might end up allocating lambdas in a loop repeatedly. Does the compiler optimise this?

No, lambdas are extremely cheap, and the compiler optimizes them heavily. See this example: https://godbolt.org/z/PV3yP-

  * Test1 compiles to a single instruction just like Test0, because the compiler is able to inline everything, including the lambda.

  * The code generated for non-inlined calls to Foo1 is nice: just a single tail call.

  * Test2 is a bit more expensive because we disallow inlining, but all we have to do is prepare a few function arguments and make the call.

  * The code generated for the lambda in Test2 (`void rtc::FunctionView<void ()>::CallVoidPtr<Test2(int)::$_1>(rtc::FunctionView<void ()>::VoidUnion)`) just prepares a function argument and makes a tail call.

So even in cases where we don't do inlining, lambdas are extremely cheap. The most expensive thing is probably the indirect branch when we call the lambda, and even that is going to be quick because of good branch prediction if you run things in a loop and always call the same lambda code in each iteration.

It does start getting expensive if you involve std::function (change to #if 1 in the example), which copies the lambda (to an internal buffer if it's small, to the heap if it's large). And if your lambda captures an expensive type by value, you'll of course end up paying for that. But lambdas themselves are both extremely low overhead and easy for the compiler to optimize.

Patch-set: 3
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>