AEC3: More restrictive use of the shadow filter output
This change tightens the conditions for when to use the shadow filter
output in order to have less switching between the main filter and the
shadow filter. This increases audio quality as some filter switches
result in audible artifacts. The shadow filter is still used when it
performs considerably better than the main filter.

Bug: webrtc:9881,chromium:896197
Change-Id: Id867eadea20a5a8645161c238d7fc7c58ff959c9