Update patch set 9
Patch Set 9:

> Patch Set 8:
> 
> > Patch Set 8: Code-Review+1
> > 
> > lgtm for webrtc.gni
> > 
> > As for the bot errors, the possible options are
> > 
> > 1) only enable pipewire if use_sysroot = true
> > 2) set enable_pipewire = false on all bots with use_sysroot = true
> > 3) install pipewire on bots and get it added to the image we use for the bots, and probably to install-build-deps.sh as well
> > 
> > A normal developer build will use sysroot, and almost all bot builds, so I think 1) is sufficient. Try changing
> > 
> > rtc_use_pipewire = is_linux
> > 
> > to
> > 
> > rtc_use_pipewire = is_linux && use_sysroot.
> 
> I'm opposed to this. As some developers (I personally never used sysroot) and distributions build against the system libraries. For me 2) or 3) works (I would prefer 3) though).

Ok. When I think of it, 2) is kind of not an option. The invariant we want to protect is that WebRTC compiles and links if install-build-deps.sh has been run; developer will be confused if we depend on some lib but don't require it as a prerequisite. In that case it's 3) that remains as the viable option.

You will have to make a CL to https://cs.chromium.org/chromium/src/build/install-build-deps.sh?q=install-build-deps that adds pipewire in such a way that WebRTC compiles even on a system that didn't have pipewire in the first place (and has use_sysroot = false). I can look into what needs to happen to add pipewire to the bot image we use.

Patch-set: 9
Reviewer: Gerrit User 5027 <5027@58829da1-049c-39fb-b951-ebdcd0984767>