Update patch set 1
Patch Set 1:

> Patch Set 1:
> OTOH, if it's a requirement, why is it done via a setter rather than a mandatory parameter in the constructor?

Good question. The answer, I think, is that the observer is also a mandatory argument to CreatePeerConnection. So a setter method is needed need to arrange the circular references.

It would make some sense with a helper CreatePeerConnectionWithMockObserver to create a peerconnection with the observer and call observer->SetPeerConnectionInterface. If there's some reasonable place in the test code to add that.

> Yes. But again, I would find it more robust to enforce the proper wiring at compile time rather than crashing if it has not been set properly.

Agreed. But this change doesn't enforce proper wiring, it just tolerates certain errors. Right?

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>