Update patch set 19
Patch Set 19:

(4 comments)

Updated review with the following changes:
* Rebased on master with too long variable name replaced by extmap_allow_mixed.
* The code has been restructured so that the SDP negotiation is properly propagated.

To be honest I think that the relation between SDP negotiation and creating audio/video streams etc is a bit messy. Which code path that is used depends on who initialized the call. In one case, RTPSender for the voice channel is created before the negotiation is complete which is why the value needs to be propagated using a chain of setters. There are also multiple copies of the same parameters. I have tried to mimic the existing design for the propagation of the new parameter.

Patch-set: 19
Reviewer: Gerrit User 7641 <7641@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0