Update patch set 1
Patch Set 1:

> The answer, I think, is that the observer is also a mandatory argument to CreatePeerConnection. So a setter method is needed need to arrange the circular references.

Fair enough :)

> It would make some sense with a helper CreatePeerConnectionWithMockObserver to create a peerconnection with the observer and call observer->SetPeerConnectionInterface. If there's some reasonable place in the test code to add that.

It's already done there, I'm going to look if it makes sense to reuse this wrapper.
https://cs.chromium.org/chromium/src/third_party/webrtc/pc/peerconnectionwrapper.cc?rcl=8e6749e0dd3fd077b5abde2fc91917dd2477f584&l=33

> But this change doesn't enforce proper wiring, it just tolerates certain errors. Right?
True. I'm still not sure it is an error: the PeerConnection is plugged in order to perform additional tests in the mock class. But maybe we don't want to exercice these tests in other situations (<- in that case maybe a simpler mock could be the answer).

Patch-set: 1
Reviewer: Gerrit User 7285 <7285@58829da1-049c-39fb-b951-ebdcd0984767>