Update patch set 4
Patch Set 4:

I thought a bit about this problem and I think we need to simplify this presubmit check. See https://webrtc-review.googlesource.com/c/src/+/107731, I think it is more actionable and robust (only check that api/ headers don't include things outside of api/, the exceptions are not enforced and left to the developer to figure out).

Patch-set: 4