Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Implementation of Java SDK support for Injectable Audio Codecs.
> 
> A few things to note:
> I added the cc files to the 'audio_jni' library. Looked reasonable to me. It already has deps on the native builtinaudio(encoder|decoder)factories.

We should create a new target and work on getting rid of the dependencies from audio_jni target.

> Should the cc files go in the sdk/android/src/jni directory, or in its pc/ subdirectory? Both seemed reasonable to me, I couldn't see the distinction.

Probably not under pc/, it is a bit strange directory. Mostly targets in peerconnection_jni are inside this directory.

> Since the Audio(Encoder|Decoder)Factory objects are reference counted, and each object knows its own reference count, passing the objects from Java to C in this way looks a bit wonky. The receiving C side decreases the reference count on behalf of the Java side. This is so that the Java side doesn't have to care about the objects' lifetime. On the other hand, we don't want to fiddle with the factories once they have been created.

Yes, this is ugly but this is how it is going to be with this approach with factory factories.

> WDYT?

Patch-set: 1
Reviewer: Gerrit User 5112 <5112@58829da1-049c-39fb-b951-ebdcd0984767>