Removing ChannelSendProxy, replacing with interface.
The main motivation for ChannelSendProxy is to allow mocking of
ChannelSend. By introducing ChannelSendInterface, this interface
can be mocked instead and ChannelSend can implement the interface
directly.

This CL reduces the amount of indirection and superfluous classes. And
prepares for increasing the integration between the generic bandwidth
estimation and allocation functionality and the audio stream control.

Bug: webrtc:9718
Change-Id: I4321181f392b7508fbaf5aec096f66a0df2d1264