Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Implementation of Java SDK support for Injectable Audio Codecs.
> > 
> > A few things to note:
> > I added the cc files to the 'audio_jni' library. Looked reasonable to me. It already has deps on the native builtinaudio(encoder|decoder)factories.
> 
> We should create a new target and work on getting rid of the dependencies from audio_jni target.
> 
> > Should the cc files go in the sdk/android/src/jni directory, or in its pc/ subdirectory? Both seemed reasonable to me, I couldn't see the distinction.
> 
> Probably not under pc/, it is a bit strange directory. Mostly targets in peerconnection_jni are inside this directory.
> 
> > Since the Audio(Encoder|Decoder)Factory objects are reference counted, and each object knows its own reference count, passing the objects from Java to C in this way looks a bit wonky. The receiving C side decreases the reference count on behalf of the Java side. This is so that the Java side doesn't have to care about the objects' lifetime. On the other hand, we don't want to fiddle with the factories once they have been created.
> 
> Yes, this is ugly but this is how it is going to be with this approach with factory factories.

Ah, I just found out what AudioProcessingFactory does.
It creates a rtc::RefCountedObject but does not assign it to a rtc::scoped_refptr. The object is directly returned to Java. This means it has reference count 0 when it gets to the Java side. When it's passed back to C++ and wrapped by a rtc::scoped_refptr the reference counter is incremented to 1, meaning that that's the only reference.

With my current implementation the object reaches the Java side with reference count 1 which is undesirable.

I'll see if I can do it the same way as AudioProcessingFactory.

I'll update the code to address the other comments as well.

Patch-set: 1
Reviewer: Gerrit User 7791 <7791@58829da1-049c-39fb-b951-ebdcd0984767>