Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> Tom, do you know what could be wrong with the test? (regarding to RPATH/RUNPATH when the sysroot is enabled).

That is intentional.  The sysroot is only used to make the build hermetic.  The RPATH is not set because otherwise we'd have to ship the sysroot to users along with Chrome.

If you want to use pipewire on systems that have the library installed, you'll have to dlopen the library.

Patch-set: 13