Add thread safety annotations on APM private submodules
It is not possible to use the locking annotations on members of substructs,
so the entire struct has to be guarded. This splits the submodules into
those guarded by crit_capture_, and those guarded by crit_render_.

For the lock annotations to also work for EchoControl, which has one
thread-safe function, this CL introduces a non-checked helper function for
that one function call.

Bug: webrtc:9929
Change-Id: Ib2ce111269956de1c07b5ee616e85508623f56a0