Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > PTAL. How do you suggest that I test the turnserver refactoring? I'd imagine one could do something with netcat and some shell scripting, but it doesn't sound that fun.
> 
> Add a simple unit test for TurnFileAuth?

Hmm. Maybe one could arrange it as

1. A helper function populating the map from an istream. Should be easy to unit test.

2. Let TurnFileAuth take the mapping as a constructor argument.

3. Main function would then open the file, call helper, and construct the TurnFileAuth.

To write a unit test for the helper function, does it have to go in a separate file? Or are there any easy way to write a test for an internal helper? 

One option might be to write an ad-hoc self test into main of turnserver_main.cc, invoked with ./turnserver --test or the like. But that seems contrary to the conventions for all other tests.

Unit-testing the population of the mapping would be adequate for testing this refactoring, I think, but to get confidence that the turnserver stays in working shape, we would really need some sort of end-to-end test.

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0