Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> > Patch Set 13:
> > 
> > Tom, do you know what could be wrong with the test? (regarding to RPATH/RUNPATH when the sysroot is enabled).
> 
> That is intentional.  The sysroot is only used to make the build hermetic.  The RPATH is not set because otherwise we'd have to ship the sysroot to users along with Chrome.
> 
> If you want to use pipewire on systems that have the library installed, you'll have to dlopen the library.

Tom: ok, but if we can't put it in install-build-deps.sh, it appears we can't get it on the bot images either. Umm, wait, since the bots are Ubuntu and libpipewire isn't available on Ubuntu, this means that by definition the bots can't run this code anyway.

Ok, Tomáš, the easiest way out here is to simply set rtc_use_pipewire=false by default, and then you have to override it with gn gen --args="rtc_use_pipewire=true" out/Whatever when you build WebRTC on systems that have pipewire. There might be fancier ways where you detect Wayland in gn, or even if the package is installed on that particular system, but I'm not aware of that being done elsewhere in the code. Setting it to false should allow you to get the patch in at least.

Patch-set: 13
Reviewer: Gerrit User 5027 <5027@58829da1-049c-39fb-b951-ebdcd0984767>