Update patch set 2
Patch Set 2:

> > Hmm. Maybe one could arrange it as
> > 
> > 1. A helper function populating the map from an istream. Should be easy to unit test.
> > 
> > 2. Let TurnFileAuth take the mapping as a constructor argument.
> > 
> > 3. Main function would then open the file, call helper, and construct the TurnFileAuth.
> 
> SGTM. It would solve the `const` issue, too.

Done. This was the first unittest under examples/, as far as I could find. So needed some plumbing in the build files, with a new executable examples_unittests. Is that right? What's needed to get try bots to run it?

Patch-set: 2
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>