Update patch set 6
Patch Set 6:

(4 comments)

Thanks!
The change for the annotations looks good. However, the change for the thread-safe function for the EchoControl I don't really follow why it should be necessary. 
Afaics it is called from within a function that already can only be called while holding the render lock so it should be possible to apply the annotation for that.

Patch-set: 6
Reviewer: Gerrit User 5125 <5125@58829da1-049c-39fb-b951-ebdcd0984767>