Add thread safety annotations to APM private submodules
It is not possible [1] to use the locking annotations on members of
substructs, so the entire struct has to be guarded. This splits the
submodules into those guarded by crit_capture_, and those guarded by
crit_render_.
(We could also make all submodules direct members of APM, but that
seems disorganized.)

For the lock annotations to also work for EchoControl, which has one
thread-safe function, this CL introduces a non-checked helper function
for that one function call.

[1] https://clang.llvm.org/docs/ThreadSafetyAnalysis.html#lexical-scope

Bug: webrtc:9929
Change-Id: Ib2ce111269956de1c07b5ee616e85508623f56a0