Update patch set 7
Patch Set 7:

(5 comments)

> Patch Set 6:
> 
> (4 comments)
> 
> Thanks!
> The change for the annotations looks good. However, the change for the thread-safe function for the EchoControl I don't really follow why it should be necessary. 
> Afaics it is called from within a function that already can only be called while holding the render lock so it should be possible to apply the annotation for that.

Thanks for having a look! In short, the annotations are not granular enough to understand that EchoControl::AnalyzeRender is safe to use in parallel with the other interface functions. I tried to lay it out in the in-line comment. If it's not clear - let's chat offline :)

Patch-set: 7
Reviewer: Gerrit User 5119 <5119@58829da1-049c-39fb-b951-ebdcd0984767>