Update patch set 7
Patch Set 7:

(2 comments)

Thanks!
I think it would be good to see a design or a CL that illustrates what the end-goal is intended to be.

Taken as a standalone CL, my feeling is that this CL adds thread confidence by locking down the code with annotations but also adds code complexity. No security is added for the current code at hand but the annotations definitely are beneficial to make future code changes more secure.

Patch-set: 7
Reviewer: Gerrit User 5125 <5125@58829da1-049c-39fb-b951-ebdcd0984767>