Update patch set 3
Patch Set 3:

PTAL on the second attempt of a workaround.

This version passes all the syncronization tests. The issue was that SetLastRtpTimestamp in these tests can be called faster or slower than the expected delay, but still at a somewhat constant rate.

The idea of this change is to generate the tiemstamps close to current time, as was before, but maintain constant difference between consecutive calls.

I've tested it locally, it doesn't work as good as fixing the first rtp/ntp timestamp sample in video_loopback, but now absolute majority of capture timestamp jumps are as needed - 20ms, instead of 19-21 as was before.

Patch-set: 3
Reviewer: Gerrit User 5019 <5019@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5399 <5399@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5117 <5117@58829da1-049c-39fb-b951-ebdcd0984767>