Update patch set 3
Patch Set 3:

I agree with Karl, Maybe deterministic way is better. Like repeating current 4-byte counter all the way in payload, maybe xored with some deterministicaly generated pseudorandom data.

Or, add the counter, as you did before in the first 4-bytes and then use it as a seed to generate the rest of the data randomly, like you are doing now. Or generate the data, but add some kind of checksum at the beginning/end.

This way we can actually check that received data is correct in the fake decoder.

Patch-set: 3
Reviewer: Gerrit User 5117 <5117@58829da1-049c-39fb-b951-ebdcd0984767>