Update patch set 4
Patch Set 4:

(2 comments)

> Patch Set 4:
> Since the manual reset seems to be such an exception and follows a slightly different code path. I think it'd make sense to have that as a separate class, like ManualResetEvent,

Either that, or an enum or magic constant to the constructor. But since type of event needed is a very static property, I'd lean towards making it a separate class, and not allow selecting manual reset or not at run time.

Patch-set: 4
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>