Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I agree with Karl, Maybe deterministic way is better. Like repeating current 4-byte counter all the way in payload, maybe xored with some deterministicaly generated pseudorandom data.
> 
> Or, add the counter, as you did before in the first 4-bytes and then use it as a seed to generate the rest of the data randomly, like you are doing now. Or generate the data, but add some kind of checksum at the beginning/end.
> 
> This way we can actually check that received data is correct in the fake decoder.

+1. If you need data that looks random in a test, it's almost always better to use something pseudorandom that's the same every time you run the test. Even if the specific case we're discussing here actually has the property that it's data agnostic, you have nothing to lose by using pseudorandom data.

And if you want different random data for each test run in order to flush out bugs over time, it's still better to use pseudorandom data but with a truly random seed, which the tests prints out when it runs, so that you can hard-code that seed when debugging the things found by the test. This is exactly how our fuzzers work.

Patch-set: 3
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>