Update patch set 7
Patch Set 7:

(2 comments)

> Patch Set 7:
> 
> Looks very good now. And if I'm not mistaken, you've removed all the Windows specific stuff---meaning that we should be able to test this on other platforms too.
> 
> The next step would be to add tests, I think (or more likely repurpose existing tests), but it would be good to get Tommi's questions answered first. That is, whether there is something similar to Apple's GCD that can be used (I believe you answered no?), and whether the libevent implementation would work.

I'll fix up the last few suggestions. As for tests, are there tests already for the thread queuing mechanism for other platforms? The issue is there are no try-bots for WinUWP (yet), and while I can test something locally for UWP it might be best if it's possible to enable this queue to be using a test meant for a different target. That way it gets coverage until such a date where full WinUWP try-bots can be added. I'm a bit uncertain though how to make that happen.

Patch-set: 7
Reviewer: Gerrit User 6219 <6219@58829da1-049c-39fb-b951-ebdcd0984767>