Update patch set 1
Patch Set 1:

(3 comments)

This seems like going in the wrong direction to me. Preferably we should enforce that the observer is injected in the constructor, we haven't gotten far enough to do this, as things will have to be refactored around how this class interacts with webrtc::Call.

Having to handle the state of the observer not being set adds complexity. It also enforces the requirement to grab a lock on the TaskQueue, something we want to avoid. (we haven't gotten far enough to to it yes as described above)

As soon as the legacy send side congestion controller(SSCC) is removed. The plan is to merge this class and SSCC. Not sure how this change will fit that structure.

There might be other ways to achieve the goals of this CL, but I don't know what those are. Please elaborate on that.

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>