Update patch set 2
Patch Set 2:

I think one source of the problem is that RtpTransportControllerSend is used in multiple ways in Call. For one, it is passed to the constructor of AudioSendStream and VideoSendStream. There it would be an improvement to wire it up via the config structs instead, in a similar way as MediaTransport (and it could be left as null with MediaTransport. Then we also need to move construction of RtpTransportControllerSend. That's an essential step for refactored RtpTransport, which we haven't gotten to yet.

Then Call also sets up a few observer links, maybe that could be moved to the code that creates Call? For the CallStatsObserver, it's not clear to me what that is for, but looks like the opposite direction of TargettransferRate.

If you need an easy way to just tell Call whether or not RtpTransport is in use, you could add some flag to CallConfig.

Patch-set: 2
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>