Update patch set 2
Patch Set 2:

I would prefer if we can postpone creating Call until we know if we're going to use MediaTransport or not.

Then I'm not sure I understand the purpose of the OnTransportChanged callback. Does it indicate that a new transport is to be used, or that a transport is somehow reconfigured?

Patch-set: 2
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>